[KYUUBI #2571] Minimize YARN tests overhead
